+++ src/cmd/compile/internal/ssa/type.go
@@ -12,7 +12,7 @@ import "cmd/internal/obj"
 // Type instances are not guaranteed to be canonical.
 type Type interface {
 	Size() int64 // return the size in bytes
-	Alignment() int64
+	Alignment() uint8
 
 	IsBoolean() bool // is a named or unnamed boolean type
 	IsInteger() bool //  ... ditto for the others
@@ -60,7 +60,7 @@ type CompilerType struct {
 }
 
 func (t *CompilerType) Size() int64            { return t.size } // Size in bytes
-func (t *CompilerType) Alignment() int64       { return 0 }
+func (t *CompilerType) Alignment() uint8       { return 0 }
 func (t *CompilerType) IsBoolean() bool        { return false }
 func (t *CompilerType) IsInteger() bool        { return false }
 func (t *CompilerType) IsSigned() bool         { return false }
@@ -95,7 +95,7 @@ type TupleType struct {
 }
 
 func (t *TupleType) Size() int64          { panic("not implemented") }
-func (t *TupleType) Alignment() int64     { panic("not implemented") }
+func (t *TupleType) Alignment() uint8     { panic("not implemented") }
 func (t *TupleType) IsBoolean() bool      { return false }
 func (t *TupleType) IsInteger() bool      { return false }
 func (t *TupleType) IsSigned() bool       { return false }
